helper functions in helper file
